requirements.txt
htcondor
influxdb
influxdb-client
PyYAML
schedule